+++ src/NzbDrone.Core/AuthorStats/BookStatistics.cs
@@ -7,6 +7,8 @@ namespace NzbDrone.Core.AuthorStats
         public int AuthorId { get; set; }
         public int BookId { get; set; }
         public int BookFileCount { get; set; }
+        public int EbookFileCount { get; set; }
+        public int AudiobookFileCount { get; set; }
         public int BookCount { get; set; }
         public int AvailableBookCount { get; set; }
         public int TotalBookCount { get; set; }