+++ src/Readarr.Api.V1/Books/BookStatisticsResource.cs
@@ -5,6 +5,8 @@ namespace Readarr.Api.V1.Books
     public class BookStatisticsResource
     {
         public int BookFileCount { get; set; }
+        public int EbookFileCount { get; set; }
+        public int AudiobookFileCount { get; set; }
         public int BookCount { get; set; }
         public int TotalBookCount { get; set; }
         public long SizeOnDisk { get; set; }
@@ -35,6 +37,8 @@ namespace Readarr.Api.V1.Books
             return new BookStatisticsResource
             {
                 BookFileCount = model.BookFileCount,
+                EbookFileCount = model.EbookFileCount,
+                AudiobookFileCount = model.AudiobookFileCount,
                 BookCount = model.BookCount,
                 SizeOnDisk = model.SizeOnDisk,
                 TotalBookCount = model.TotalBookCount